app/requirements.txt
Django
djangorestframework
sqlalchemy
psycopg2
